fix(MeaSys): Fix C4244 warning in RsMeasure (1 fix)
- Line 109: double to int conversion for iMeasTime calculation
- Added static_cast<int>() for dScanTime * 1000 expression
- Ensures type safety and eliminates possible data loss warning

+++ SRC/MeaSys/RsMeasure.cpp
@@ -106,7 +106,7 @@ BOOL CRsMeasure::Measure(const MEAS_PROG_INFO* pMeasProgInfo)
 
 	// Jn
 	iRange = atoi(m_pszRange[pMeasProgInfo->ScanParams._RS.lStartIdx]);
-	iMeasTime = pMeasProgInfo->ScanParams._RS.dScanTime * 1000;
+	iMeasTime = static_cast<int>(pMeasProgInfo->ScanParams._RS.dScanTime * 1000);
 
 	if ( !m_pRsHead->MeasStart() ) {
 		MoveProbeHeadLowerPos(FALSE);
